RegistryOpsSpec eliminated
